Ajuste na migration 13

+++ src/main/resources/db/migration/V13__Create_Table_Participacao_Promocao.sql
@@ -1,11 +1,11 @@
 CREATE TABLE milhas.participacao_promocao (
-    id SERIAL PRIMARY KEY,
-    usuario_id BIGINT NOT NULL,
-    promocao_id BIGINT NOT NULL,
-    data_adesao TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+      id BIGSERIAL PRIMARY KEY,
+      usuario_id BIGINT NOT NULL,
+      promocao_id BIGINT NOT NULL,
+      data_adesao TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
 
-    CONSTRAINT fk_participacao_usuario FOREIGN KEY (usuario_id) REFERENCES milhas.usuario(id),
-    CONSTRAINT fk_participacao_promocao FOREIGN KEY (promocao_id) REFERENCES milhas.promocao(id),
+      CONSTRAINT fk_participacao_usuario FOREIGN KEY (usuario_id) REFERENCES milhas.usuario(id),
+      CONSTRAINT fk_participacao_promocao FOREIGN KEY (promocao_id) REFERENCES milhas.promocao(id),
 
-    CONSTRAINT uk_usuario_promocao UNIQUE (usuario_id, promocao_id)
+      CONSTRAINT uk_usuario_promocao UNIQUE (usuario_id, promocao_id)
 );
